% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_building.r
\name{create_bizkaia_cad_building_age}
\alias{create_bizkaia_cad_building_age}
\title{Using the information of bizkaia cadastre, generate map with the building age information}
\usage{
create_bizkaia_cad_building_age(city)
}
\description{
Using the information of bizkaia cadastre, generate map with the building age information
}
\examples{
create_bizkaia_cad_building_age("Bilbao")
}
