% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_building.r
\name{create_spanish_cad_building_age}
\alias{create_spanish_cad_building_age}
\title{Using the information of cadastre, generate map with the building age information}
\usage{
create_spanish_cad_building_age(city)
}
\description{
Using the information of cadastre, generate map with the building age information
}
\examples{
create_spanish_cad_building_age(String city)
}
